+++ src/app/applicant/page.tsx
@@ -12,6 +12,7 @@ import {
   ArrowRight,
 } from "lucide-react";
 import { checkApplicationStatus } from "@/lib/redux/api/applicantApi";
+import { RecentActivityCard } from "./RecentActivityCard";
 
 // Inline ApplicationTimeline
 interface TimelineItemProps {
@@ -84,7 +85,7 @@ function ApplicationTimeline({ application_id }: { application_id: string }) {
   useEffect(() => {
     async function fetchTimeline() {
       try {
-        const response = await checkApplicationStatus(application_id);
+        const response = await checkApplicationStatus();
         setTimelineItems(response.data.timeline || []);
       } catch (err) {
         if (err instanceof Error) {
@@ -122,67 +123,7 @@ function ApplicationTimeline({ application_id }: { application_id: string }) {
   );
 }
 
-export interface ActivityType {
-  type: string;
-  title: string;
-  date?: string;
-}
 
-export function RecentActivityCard({
-  application_id,
-}: {
-  application_id: string;
-}) {
-  const [activities, setActivities] = useState<ActivityType[]>([]); // ...existing code...
-  const [loading, setLoading] = useState<boolean>(true); // ...existing code...
-  const [error, setError] = useState<string | null>(null); // ...existing code...
-
-  useEffect(() => {
-    async function fetchActivities() {
-      try {
-        const response = await checkApplicationStatus(application_id);
-        setActivities(response.data.activities || []);
-      } catch (err) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError("Failed to fetch activities");
-        }
-      } finally {
-        setLoading(false);
-      }
-    }
-    if (application_id) fetchActivities();
-  }, [application_id]);
-
-  if (loading) return <p>Loading activities...</p>;
-  if (error) return <p className="text-red-500">{error}</p>;
-
-  return (
-    <Card className="shadow-sm">
-      <CardHeader>
-        <CardTitle className="text-xl font-semibold text-gray-800">
-          Recent Activity
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-4">
-        {activities.map((act, idx) => (
-          <div key={idx} className="flex items-center gap-3">
-            {act.type === "submitted" ? (
-              <CheckCircle className="w-5 h-5 text-green-500" />
-            ) : (
-              <Calendar className="w-5 h-5 text-gray-500" />
-            )}
-            <div>
-              <p className="font-medium text-gray-700">{act.title}</p>
-              <p className="text-sm text-gray-500">{act.date}</p>
-            </div>
-          </div>
-        ))}
-      </CardContent>
-    </Card>
-  );
-}
 
 // Inline ImportantUpdatesCard
 function ImportantUpdatesCard() {
